nquery add dump user 	modified:   apps/nquery/src/main.c 	modified:   apps/ntrackd/src/main.c 	modified:   apps/ntrackd/src/nt_netlink.c 	modified:   apps/ntrackd/src/nt_unotify.c 	modified:   drivers/nos_track/src/ntrack_conf.c 	modified:   drivers/nos_track/src/ntrack_hooks.c 	modified:   include/src/ntrack_auth.h 	modified:   include/src/ntrack_log.h 	modified:   include/src/ntrack_msg.h

+++ apps/ntrackd/src/nt_netlink.c
@@ -78,7 +78,7 @@ int nt_nl_xmit(void *data)
 	message.hdr.nlmsg_pid = getpid();
 
 	memcpy(NLMSG_DATA(&message), data, strlen(data));
-	nt_info("message send to kernel: %d bytes.\n", strlen(data));
+	nt_info("message send to kernel: %d bytes.\n", (int)strlen(data));
 	ret = sendto(nl_sock, 
 		&message, 
 		message.hdr.nlmsg_len, 0,
